Add files via upload
Using data from the U.S Bureau of Labor of Statistics, I  analyzed unemployement statistics for the years 2018, 2019, and 2020.
The Covid-19 pandemic began right around the new year of the year 2020. Using unemployment statsitics from the U.S. Bureau of Labor Statistics, the goal is to compare unemployment statistics two years prior to the onset of the Covid-19 pandemic and compare that to unemployment statistics during the year of 2020 when the pandemic was at its peak and determine if there is a significant difference in unemployment pre and post pandemic.

By extracting the data and using Power Pivot, Pivot Tables and a two sided t-test, I was able to answer the following questions:

1) Is there a statisical significant difference in unemployment pre and post onset of the pandemic?
3) Which states suffered from the highest unmeployemnt post pandemic, and which states had the lowest?
4) What is the average unemployment across the country across any number of different quarters for the years 2018-2020?
---
This graphic provides the extracted data of unemployment stattisitcs for the months of October 2017 to the months of December 2020.

<img width="720" alt="Screen Shot 2021-11-26 at 5 24 43 PM" src="https://user-images.githubusercontent.com/95102232/143660040-5887eb0a-6cd2-480e-868b-9693d26fb562.png">

---
This graphic illustrates the average unemployment rate for the years 2018, 2019, 2020, and the average of 2018 and 2019.

<img width="812" alt="Screen Shot 2021-11-26 at 6 11 40 PM" src="https://user-images.githubusercontent.com/95102232/143660525-95ec6c0b-5618-44d1-bd3a-80d1f97af1d0.png">

---
Using pivot tables, slicers, and timelines, I was able to determine the states that had the highest and lowest average rate of unemployemnt between 2018 and 2019 and in 2020. This image also uses a timeline to depict unemployment across different quarters throughout these three years.

<img width="659" alt="Screen Shot 2021-11-26 at 6 32 58 PM" src="https://user-images.githubusercontent.com/95102232/143661092-a7752540-4548-4121-a6b5-4aa0f378e82f.png">

---
These are two heat maps comparing average rates of unemployment between 2018 and 2019 and in 2020.
<img width="1066" alt="Screen Shot 2021-11-26 at 6 43 06 PM" src="https://user-images.githubusercontent.com/95102232/143661358-c9ebf2d6-c8d4-4a87-a3b6-d9a8685224ba.png">

---
This shows the two sample t-Test which provides 99% confidence that the difference in unemployement between 2018-2019 and 2020
are not due to chance alone.

<img width="501" alt="Screen Shot 2021-11-26 at 6 43 59 PM" src="https://user-images.githubusercontent.com/95102232/143661453-90ed5842-c34f-40f1-ac24-aeaabc55e1ce.png">